Fix bugs related to specifying slider min and max differing from the default 0 and 100.
Per update sent to vpusers list by Bruce Sherwood:

Attached is a replacement for the file site-packages/vis/controls.py.
The attached file fixes bugs in specifying slider min and max
different from the default values of 0 and 100. This fix will of
course be included in later releases of VPython. I thank Frank
Noschese for reporting the bug.